+++ src/components/Gallery.js
@@ -44,8 +44,7 @@ export class Gallery extends React.PureComponent {
           <div
             className="str-chat__gallery-placeholder"
             style={{
-              background: `url(${images[3].image_url}) top left no-repeat`,
-              backgroundSize: 'cover',
+              backgroundImage: `url(${images[3].image_url})`,
             }}
             onClick={() => this.toggleModal(3)}
           >
@@ -63,7 +62,7 @@ export class Gallery extends React.PureComponent {
         /> */}
         <ModalGateway>
           {this.state.modalIsOpen ? (
-            <Modal onClose={this.toggleModal}>
+            <Modal onClose={this.toggleModal} closeOnBackdropClick={true}>
               <Carousel
                 views={formattedArray}
                 currentIndex={this.state.currentIndex}