+++ src/components/MessageInputLarge.js
@@ -22,6 +22,8 @@ export class MessageInputLarge extends PureComponent {
     focus: PropTypes.bool.isRequired,
     /** Grow the textarea while you're typing */
     grow: PropTypes.bool.isRequired,
+    /** Specify the max amount of rows the textarea is able to grow */
+    maxRows: PropTypes.number.isRequired,
     /** Make the textarea disabled */
     disabled: PropTypes.bool,
     /** @see See [MessageInput](https://getstream.github.io/stream-chat-react/#messageinput) for doc */
@@ -132,7 +134,7 @@ export class MessageInputLarge extends PureComponent {
   render() {
     const SendButton = this.props.SendButton;
     return (
-      <div style={{ position: 'relative', zIndex: 0, width: '100%' }}>
+      <div className="str-chat__input-large">
         <ImageDropzone
           accept={this.props.acceptedFiles}
           multiple={this.props.multipleUploads}
@@ -156,6 +158,7 @@ export class MessageInputLarge extends PureComponent {
                 onSelectItem={this.props.onSelectItem}
                 value={this.props.text}
                 rows={1}
+                maxRows={this.props.maxRows}
                 placeholder="Type your message"
                 onPaste={this.props.onPaste}
                 grow={this.props.grow}