+++ src/components/MessageSimple.js
@@ -117,6 +117,11 @@ export class MessageSimple extends PureComponent {
      * @param user Target user object
      */
     onMentionsClickMessage: PropTypes.func,
+    /**
+     * Additional props for underlying MessageInput component.
+     * Available props - https://getstream.github.io/stream-chat-react/#messageinput
+     * */
+    additionalMessageInputProps: PropTypes.object,
   };
 
   static defaultProps = {
@@ -493,6 +498,7 @@ export class MessageSimple extends PureComponent {
               message={message}
               clearEditingState={clearEditingState}
               updateMessage={updateMessage}
+              {...this.props.additionalMessageInputProps}
             />
           </Modal>
         )}