+++ types/index.d.ts
@@ -51,6 +51,7 @@ export interface ChannelContextValue extends ChatContextValue {
   acceptedFiles?: string[];
   maxNumberOfFiles?: number;
   sendMessage?(message: Client.Message): void;
+  editMessage?(updatedMessage: Client.Message): void;
   /** Via Context: The function to update a message, handled by the Channel component */
   updateMessage?(
     updatedMessage: Client.MessageResponse,
@@ -108,6 +109,17 @@ export interface ChannelProps extends ChatContextValue {
   onMentionsClick?(e: React.MouseEvent, user: Client.UserResponse): void;
   /** Function to be called when hovering over a @mention. Function has access to the DOM event and the target user object */
   onMentionsHover?(e: React.MouseEvent, user: Client.UserResponse): void;
+
+  /** Override send message request (Advanced usage only) */
+  doSendMessageRequest?(
+    channelId: string,
+    message: Client.Message,
+  ): Promise<Client.MessageResponse> | void;
+  /** Override update(edit) message request (Advanced usage only) */
+  doUpdateMessageRequest?(
+    channelId: string,
+    updatedMessage: Client.Message,
+  ): Promise<Client.MessageResponse> | void;
 }
 
 export interface ChannelListProps extends ChatContextValue {
@@ -137,6 +149,14 @@ export interface ChannelListProps extends ChatContextValue {
     thisArg: React.Component<ChannelListProps>,
     e: Client.Event<Client.ChannelUpdatedEvent>,
   ): any;
+  onChannelDeleted?(
+    thisArg: React.Component<ChannelListProps>,
+    e: Client.Event<Client.ChannelDeletedEvent>,
+  ): void;
+  onChannelTruncated?(
+    thisArg: React.Component<ChannelListProps>,
+    e: Client.Event<Client.ChannelTruncatedEvent>,
+  ): void;
   /** Object containing query filters */
   filters: object;
   /** Object containing query options */
@@ -234,6 +254,11 @@ export interface EmptyStateIndicatorProps {
   listType: string;
 }
 
+export interface SendButtonProps {
+  /** Function that gets triggered on click */
+  sendMessage?(message: Client.Message): void;
+}
+
 export interface MessageListProps extends ChannelContextValue {
   /** Typing indicator component to render  */
   TypingIndicator?: React.ElementType<TypingIndicatorProps>;
@@ -255,6 +280,7 @@ export interface MessageListProps extends ChannelContextValue {
   getFlagMessageErrorNotification?(message: MessageResponse): string;
   getMuteUserSuccessNotification?(message: MessageResponse): string;
   getMuteUserErrorNotification?(message: MessageResponse): string;
+  additionalMessageInputProps?: object;
 }
 
 export interface ChannelHeaderProps extends ChannelContextValue {
@@ -271,12 +297,17 @@ export interface MessageInputProps {
   disabled?: boolean;
   /** Grow the textarea while you're typing */
   grow?: boolean;
+  /** Max number of rows the textarea is allowed to grow */
+  maxRows?: number;
 
   /** The parent message object when replying on a thread */
   parent?: Client.MessageResponse | null;
 
   /** The component handling how the input is rendered */
   Input?: React.ElementType<MessageInputUIComponentProps>;
+
+  /** Change the SendButton component */
+  SendButton?: React.ElementType<SendButtonProps>;
 
   /** Override image upload request */
   doImageUploadRequest?(file: object, channel: Client.Channel): void;
@@ -396,6 +427,7 @@ export interface MessageProps {
     message: Client.MessageResponse,
     event: React.SyntheticEvent,
   ): void;
+  additionalMessageInputProps?: object;
 }
 
 export interface MessageUIComponentProps extends MessageProps {
@@ -423,6 +455,7 @@ export interface MessageUIComponentProps extends MessageProps {
   ): void;
   channelConfig?: object;
   threadList?: boolean;
+  additionalMessageInputProps?: object;
 }
 
 export interface ThreadProps extends ChannelContextValue {
@@ -430,6 +463,9 @@ export interface ThreadProps extends ChannelContextValue {
   fullWidth?: boolean;
   /** Make input focus on mounting thread */
   autoFocus?: boolean;
+  additionalParentMessageProps?: object;
+  additionalMessageListProps?: object;
+  additionalMessageInputProps?: object;
 }
 
 export interface TypingIndicatorProps {