+++ CHANGELOG.md
@@ -1,5 +1,36 @@
 # Changelog
 
+## [0.7.17](https://github.com/GetStream/stream-chat-react/releases/tag/v0.7.17) 2019-01-02
+
+- Added `maxRows` props to MessageInput component
+
+## [0.7.16](https://github.com/GetStream/stream-chat-react/releases/tag/v0.7.16) 2019-01-02
+
+- Removed inline styles from multiple locations
+- Exporting new component `ChatAutoComplete` (Advanced usage only)
+
+## [0.7.15](https://github.com/GetStream/stream-chat-react/releases/tag/v0.7.15) 2019-12-30
+
+- Added the following props on the `Thread` component so the underlying MessageList, MessageInput and Message components can be customized using props:
+  - `additionalParentMessageProps`
+  - `additionalMessageListProps`
+  - `additionalMessageInputProps`
+- Added the following props to the `Channel` component:
+  - `doUpdateMessageRequest` to override the update(edit) message request (Advanced usage only)
+  - `doSendMessageRequest` to override the send message request (Advanced usage only)
+
+## [0.7.13](https://github.com/GetStream/stream-chat-react/releases/tag/v0.7.13) 2019-12-03
+
+- Handling and updating channel list on `channel.truncated` and `channel.deleted` event.
+
+## [0.7.12](https://github.com/GetStream/stream-chat-react/releases/tag/v0.7.12) 2019-11-22
+
+- Adding prop `MessageSystem` to customize system messages
+
+## [0.7.11](https://github.com/GetStream/stream-chat-react/releases/tag/v0.7.11) 2019-11-05
+
+- Fixed z-index issue on MessageInputLarge component https://github.com/GetStream/stream-chat-react/commit/f78b0bf6566fe587da62a8162ab5f1b3d799a10e
+
 ## [0.7.10](https://github.com/GetStream/stream-chat-react/releases/tag/v0.7.10) 2019-10-16
 
 - Added `customActiveChannel` prop on `ChannelList` to specify a custom channel to be moved to the top and set to active upon mounting.